% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.AddNewField.R
\docType{data}
\name{MDataUpdater.AddNewField}
\alias{MDataUpdater.AddNewField}
\title{MDataUpdater.AddNewField}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.AddNewField
}
\description{
MDataUpdater.AddNewField
}
\keyword{datasets}

