% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.Custom.R
\docType{data}
\name{MDataUpdater.Custom}
\alias{MDataUpdater.Custom}
\title{MDataUpdater.Custom}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.Custom
}
\description{
MDataUpdater.Custom
}
\keyword{datasets}

