% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.CustomTaxonomy.R
\docType{data}
\name{MDataUpdater.CustomTaxonomy}
\alias{MDataUpdater.CustomTaxonomy}
\title{MDataUpdater.CustomTaxonomy}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.CustomTaxonomy
}
\description{
MDataUpdater.CustomTaxonomy
}
\keyword{datasets}

