% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.ExtractByTime.R
\docType{data}
\name{MDataUpdater.ExtractByTime}
\alias{MDataUpdater.ExtractByTime}
\title{MDataUpdater.ExtractByTime}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.ExtractByTime
}
\description{
MDataUpdater.ExtractByTime
}
\keyword{datasets}

