% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.FRED.R
\docType{data}
\name{MDataUpdater.FRED}
\alias{MDataUpdater.FRED}
\title{MDataUpdater.FRED}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.FRED
}
\description{
MDataUpdater.FRED
}
\keyword{datasets}

