% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.GoogleFinance.News.R
\docType{data}
\name{MDataUpdater.GoogleFinance.News}
\alias{MDataUpdater.GoogleFinance.News}
\title{MDataUpdater.GoogleFinance.News}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.GoogleFinance.News
}
\description{
MDataUpdater.GoogleFinance.News
}
\keyword{datasets}

