% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.InvestFunds.FundList.R
\docType{data}
\name{MDataUpdater.InvestFunds.FundList}
\alias{MDataUpdater.InvestFunds.FundList}
\title{MDataUpdater.InvestFunds.FundHistory}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.InvestFunds.FundList
}
\description{
MDataUpdater.InvestFunds.FundHistory
}
\keyword{datasets}

