% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.MultipleMerge.R
\docType{data}
\name{MDataUpdater.MultipleMerge}
\alias{MDataUpdater.MultipleMerge}
\title{MDataUpdater.MultipleMerge}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.MultipleMerge
}
\description{
MDataUpdater.MultipleMerge
}
\keyword{datasets}

