% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.Universe.Russian.Equities.R
\docType{data}
\name{MDataUpdater.Universe.Russian.Equities}
\alias{MDataUpdater.Universe.Russian.Equities}
\title{MDataUpdater.Universe.Russian.Equities}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.Universe.Russian.Equities
}
\description{
MDataUpdater.Universe.Russian.Equities
}
\keyword{datasets}

