% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.Universe.YF.CorpKeyStat.R
\docType{data}
\name{MDataUpdater.Universe.YF.CorpKeyStat}
\alias{MDataUpdater.Universe.YF.CorpKeyStat}
\title{MDataUpdater.Universe.YF.CorpKeyStat}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.Universe.YF.CorpKeyStat
}
\description{
MDataUpdater.Universe.YF.CorpKeyStat
}
\keyword{datasets}

