% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.Universe.YF.IndexComponents.R
\docType{data}
\name{MDataUpdater.Universe.YF.IndexComponents}
\alias{MDataUpdater.Universe.YF.IndexComponents}
\title{MDataUpdater.Universe.YF.IndexComponents}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.Universe.YF.IndexComponents
}
\description{
MDataUpdater.Universe.YF.IndexComponents
}
\keyword{datasets}

