% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDataUpdater.YF.CompanyProfile.R
\docType{data}
\name{MDataUpdater.YF.CompanyProfile}
\alias{MDataUpdater.YF.CompanyProfile}
\title{MDataUpdater.YF.CompanyProfile}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.YF.CompanyProfile
}
\description{
MDataUpdater.YF.CompanyProfile
}
\keyword{datasets}

