% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketDataUpdater.YF.fundamentals.R
\docType{data}
\name{MDataUpdater.YF.fundamentals}
\alias{MDataUpdater.YF.fundamentals}
\title{MDataUpdater.YF.fundamentals}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MDataUpdater.YF.fundamentals
}
\description{
MDataUpdater.YF.fundamentals
}
\keyword{datasets}

